more test, utilities
